Add class Proposal::Base
