+++ src/lib/y2storage/guided_proposal.rb
@@ -19,10 +19,7 @@
 # To contact SUSE LLC about this file by physical or electronic mail, you may
 # find current contact information at www.suse.com.
 
-require "yast"
 require "y2storage/proposal_settings"
-require "y2storage/storage_manager"
-require "y2storage/disk_analyzer"
 require "y2storage/exceptions"
 require "y2storage/proposal"
 
@@ -30,32 +27,23 @@ module Y2Storage
   # Class to calculate a storage proposal to install the system
   #
   # @example
-  #   proposal = Storage::GuidedProposal.new
+  #   proposal = Y2Storage::GuidedProposal.new
   #   proposal.settings.use_separate_home = true
-  #   proposal.proposed? # => false
-  #   proposal.devices   # => nil
+  #   proposal.proposed?                          # => false
+  #   proposal.devices                            # => nil
+  #   proposal.planned_devices                    # => nil
   #
-  #   proposal.propose   # Performs the calculation
+  #   proposal.propose                            # Performs the calculation
   #
-  #   proposal.proposed? # => true
-  #   proposal.devices   # Proposed layout
+  #   proposal.proposed?                          # => true
+  #   proposal.devices                            # => Proposed layout
   #   proposal.settings.use_separate_home = false # raises RuntimeError
   #
-  class GuidedProposal
-    include Yast::Logger
-
+  class GuidedProposal < Proposal::Base
     # Settings used to calculate the proposal. They cannot be altered after
     # calculating the proposal
     # @return [ProposalSettings]
     attr_reader :settings
-    # Planned devices calculated by the proposal, nil if the proposal has not
-    # been calculated yet
-    # @return [Array<Planned::Device>]
-    attr_reader :planned_devices
-    # Proposed layout of devices, nil if the proposal has not been
-    # calculated yet
-    # @return [Devicegraph]
-    attr_reader :devices
 
     class << self
       # Calculates the initial proposal
@@ -117,6 +105,8 @@ module Y2Storage
       end
     end
 
+    # Constructor
+    #
     # @param settings [ProposalSettings] if nil, default settings will be used
     # @param devicegraph [Devicegraph] starting point. If nil, the probed
     #   devicegraph will be used
@@ -124,32 +114,14 @@ module Y2Storage
     #   based on the initial {devicegraph} or will use the one in {StorageManager} if
     #   starting from probed (i.e. {devicegraph} argument is also missing)
     def initialize(settings: nil, devicegraph: nil, disk_analyzer: nil)
+      super(devicegraph: devicegraph, disk_analyzer: disk_analyzer)
       @settings = settings || ProposalSettings.new_for_current_product
-      @proposed = false
-      @initial_graph = devicegraph
-      @disk_analyzer = disk_analyzer
-
-      # Use probed devicegraph if no devicegraph is passed
-      if @initial_graph.nil?
-        @initial_graph = StorageManager.instance.y2storage_probed
-        # Use cached disk analyzer for probed devicegraph is no disk analyzer is passed
-        @disk_analyzer ||= StorageManager.instance.probed_disk_analyzer
-      end
-      # Create new disk analyzer when devicegraph is passed but not disk analyzer
-      @disk_analyzer ||= DiskAnalyzer.new(@initial_graph)
-    end
-
-    # Checks whether the proposal has already being calculated
-    # @return [Boolean]
-    def proposed?
-      @proposed
     end
 
     # Calculates the proposal
     #
     # @raise [UnexpectedCallError] if called more than once
-    # @raise [NoDiskSpaceError] if there is no enough space to
-    #           perform the installation
+    # @raise [NoDiskSpaceError] if there is no enough space to perform the installation
     def propose
       raise UnexpectedCallError if proposed?
       settings.freeze
@@ -178,20 +150,7 @@ module Y2Storage
       raise exception
     end
 
-    # A proposal is failed when it has not devices after being proposed
-    #
-    # @return [Boolean] true if proposed and has not devices; false otherwise
-    def failed?
-      proposed? && devices.nil?
-    end
-
-  protected
-
-    # Disk analyzer used to analyze the initial devicegraph
-    # @return [DiskAnalyzer]
-    attr_reader :disk_analyzer
-    # @return [Devicegraph]
-    attr_reader :initial_graph
+  private
 
     # @return [Array<Planned::Device>]
     def planned_devices_list(target)
@@ -213,10 +172,10 @@ module Y2Storage
       @space_maker ||= Proposal::SpaceMaker.new(disk_analyzer, populated_settings)
     end
 
-    # Copy of #initial_graph without all the partitions that must be wiped out
+    # Copy of #initial_devicegraph without all the partitions that must be wiped out
     # according to the settings
     def clean_graph
-      @clean_graph ||= space_maker.delete_unwanted_partitions(initial_graph)
+      @clean_graph ||= space_maker.delete_unwanted_partitions(initial_devicegraph)
     end
 
     # Copy of the original settings including some calculated and necessary
@@ -245,7 +204,7 @@ module Y2Storage
       return [populated_settings.root_device] if populated_settings.root_device
 
       disk_names = populated_settings.candidate_devices
-      candidate_disks = initial_graph.disk_devices.select { |d| disk_names.include?(d.name) }
+      candidate_disks = initial_devicegraph.disk_devices.select { |d| disk_names.include?(d.name) }
       candidate_disks = candidate_disks.sort_by(&:size).reverse
       candidate_disks.map(&:name)
     end
